IC_plot

Cumulative percentage of mature females against length. Fitting model for the estimation of the L50 values and its confidence interval.

Usage

IC_plot(x, y, yc, MClw, MCup, CIl, CIh, xl, yl)

Arguments

x

A vector of the X-axis values

y

A vector of the y-axis values

yc

A vector with Y values calculated from the original data

MClw

Lower value to be defined for the x-axis

MCup

Upper value to be defined for the x-axis

CIl

A table with the lower value of nonparametric confidence interval

CIh

A table with the upper value of nonparametric confidence interval

xl

A label for the X-axis

yl

A label for the y-axis

Value

It returs a graphic representation of the cumulative percentage of mature females as a function of length (points). The fitted model (black line) and its 95 percent confidence interval (gray shaded area) are included; the horizontal dashed line represents the cumulative ratio of 0.5; The vertical dashed line indicates the L50 value and the outer lines the 95 percent of confidence interval.

Examples

  if (FALSE) {
 IC_plot(x,y,yc,MClw,MCup,CIl,CIh,xl,yl)
  }