fn_Barplot

A very simple barplot is generated with no labels or marks for the X-axis, which serves as the basis for those that will be generated through the analysis, whether they are simple or compound figures

Usage

fn_Barplot(df, i, densidad, angulo)

Arguments

df

A list of matrices with frequencies values.

i

A numeric value of the ith object in the list.

densidad

a vector that specifies the shading lines density (in lines per inch). By default, it is NULL, which means no shading lines. For more details, see the help of the barplot() function

angulo

a vector that specifies the slope of shading lines. For more details, see the help of the barplot() function

Value

It returns a simple barplot that can be upgraded according to needs.

Examples