modelo

Fit the model using the nlsLM function. For more details see the function documentation: nlsLM() from minpack.lm (version 1.2-4).

Usage

modelo(x,y,mod,bb,bmin,bmax)

Arguments

x

independent variable.

y

dependent variable.

mod

A nonlinear model formula including variables and parameters. Will be coerced to a formula if necessary.

bb

A list or numeric vector of starting estimates. If par is a list, then each element must be of length 1.

bmin

A numeric vector of lower bounds on each parameter. If not given, the default lower bound for each parameter is set to -Inf.

bmax

A numeric vector of upper bounds on each parameter. If not given, the default upper bound for each parameter is set to Inf.

Value

Model settings list

Examples

  if (FALSE) {
 modelo(x,y,mod,bb,bmin,bmax)
  }